+++ crates/ryo3-reqwest/src/lib.rs
@@ -1,6 +1,5 @@
 #![doc = include_str!("../README.md")]
 mod body;
-mod cert;
 mod client;
 mod cookie;
 mod errors;
@@ -11,10 +10,10 @@ mod response_head;
 mod response_parking_lot;
 mod response_stream;
 mod rustls_provider;
+mod tls;
 mod tls_version;
 mod user_agent;
 
-pub use cert::PyCertificate;
 pub use client::RyBlockingClient;
 pub use client::RyHttpClient;
 pub use cookie::PyCookie;
@@ -24,6 +23,8 @@ pub use response_parking_lot::RyBlockingResponse;
 pub use response_parking_lot::RyResponse;
 pub use response_stream::RyBlockingResponseStream;
 pub use response_stream::RyResponseStream;
+pub use tls::PyCertificate;
+pub use tls::PyCertificateRevocationList;
 
 #[cfg(feature = "experimental-async")]
 pub use client::RyClient;
@@ -34,6 +35,7 @@ pub fn pymod_add(m: &Bound<'_, PyModule>) -> PyResult<()> {
 
     m.add_class::<PyCookie>()?;
     m.add_class::<PyCertificate>()?;
+    m.add_class::<PyCertificateRevocationList>()?;
     m.add_class::<RyHttpClient>()?;
     #[cfg(feature = "experimental-async")]
     m.add_class::<RyClient>()?;